move store to separate files. Fix lint errors
