+++ Sources/SmileID/Classes/DocumentVerification/Model/DocumentCaptureResultStore.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct DocumentCaptureResultStore {
+    var allFiles: [URL]
+    var documentFront: URL
+    var docmentBack: URL?
+    var selfie: URL
+    var livenessImages: [URL]
+}
