+++ Sources/SmileID/Classes/SelfieCapture/ViewModel/SelfieCaptureResultStore.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct SelfieCaptureResultStore {
+    var allFiles: [URL]
+    var selfie: URL
+    var livenessImages: [URL]
+}
